requirements.txt
# Core dependencies
torch>=1.13.0
torchvision>=0.14.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
pillow>=8.3.0
scipy>=1.7.0
scikit-learn>=1.0.0
tqdm>=4.62.0

# EEG and Brain decoding
braindecode>=0.7.0
torchinfo>=1.7.0

# Computer Vision and ML
clip-by-openai
diffusers>=0.21.0
transformers>=4.21.0
huggingface_hub>=0.16.0
accelerate>=0.20.0
ftfy>=6.1.0

# Evaluation metrics
torchmetrics>=0.11.0
pytorch-fid>=0.3.0
seaborn>=0.11.0

# Utilities
pyyaml>=6.0
jupyter>=1.0.0
ipython>=7.0.0 